+++ packages/cache-redis/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ts-linq/cache-redis
 
+## 1.0.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/core@3.4.8
+
+## 1.0.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.9.0
+  - @ts-linq/core@3.4.7
+
 ## 1.0.44
 
 ### Patch Changes