+++ packages/composite-sql-logger/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ts-linq/composite-sql-logger
 
+## 12.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/core@3.4.8
+
+## 12.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.9.0
+  - @ts-linq/core@3.4.7
+
 ## 12.0.6
 
 ### Patch Changes