+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ts-linq/telemetry
 
+## 2.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.9.0
+
 ## 2.3.1
 
 ### Patch Changes